refactor: flight | refactoring name functions and files for flight feature
